Rename as Header.tsx
